move components to examples/excalidraw/components
