+++ examples/excalidraw/components/App.scss
