+++ examples/excalidraw/components/CustomFooter.tsx
