+++ examples/excalidraw/components/MobileFooter.tsx
