+++ examples/excalidraw/components/sidebar/ExampleSidebar.scss
