+++ examples/excalidraw/components/sidebar/ExampleSidebar.tsx
